requirements3.txt

absl-py=2.1.0
aiofiles=23.2.1
altair=5.4.1
annotated-types=0.7.0
anyascii=0.3.2
anyio=4.4.0
attrs=24.2.0
audioread=3.0.1
av=10.0.0
babel=2.16.0
boto3=1.35.15
botocore=1.35.15
ca-certificates=2024.7.2=h06a4308_0
cached-path=1.6.3
cachetools=5.5.0
certifi=2024.8.30
cffi=1.17.1
charset-normalizer=3.3.2
click=8.1.7
cn2an=0.5.22
coloredlogs=15.0.1
contourpy=1.2.1
ctranslate2=3.24.0
cycler=0.12.1
cython=3.0.11
dateparser=1.1.8
decorator=5.1.1
deprecated=1.2.14
distance=0.1.3
docopt=0.6.2
dtw-python=1.4.4
eng-to-ipa=0.0.2
exceptiongroup=1.2.2
fastapi=0.114.0
faster-whisper=0.9.0
ffmpy=0.4.0
filelock=3.13.4
flatbuffers=24.3.25
fonttools=4.53.1
fsspec=2024.9.0
fugashi=1.3.0
g2p-en=2.1.0
g2pkk=0.1.2
google-api-core=2.19.2
google-auth=2.34.0
google-cloud-core=2.4.1
google-cloud-storage=2.18.2
google-crc32c=1.6.0
google-resumable-media=2.7.2
googleapis-common-protos=1.65.0
gradio=3.48.0
gradio-client=0.6.1
grpcio=1.66.1
gruut=2.2.3
gruut-ipa=0.13.0
gruut-lang-de=2.0.1
gruut-lang-en=2.0.1
gruut-lang-es=2.0.1
gruut-lang-fr=2.0.2
h11=0.14.0
httpcore=1.0.5
httpx=0.27.2
huggingface-hub=0.17.3
humanfriendly=10.0
idna=3.8
importlib-metadata=8.4.0
importlib-resources=6.4.5
inflect=7.0.0
jaconv=0.4.0
jamo=0.4.1
jieba=0.42.1
jinja2=3.1.4
jmespath=1.0.1
joblib=1.4.2
jsonlines=1.2.0
jsonschema=4.23.0
jsonschema-specifications=2023.12.1
kiwisolver=1.4.7
langid=1.1.6
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
librosa=0.9.1
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.43.0
loguru=0.7.2
markdown=3.7
markdown-it-py=3.0.0
markupsafe=2.1.5
matplotlib=3.8.4
mdurl=0.1.2
mecab-python3=1.0.9
melotts=0.1.2
more-itertools=10.5.0
mpmath=1.3.0
myshell-openvoice=0.0.0=dev_0
narwhals=1.6.4
ncurses=6.4=h6a678d5_0
networkx=2.8.8
nltk=3.9.1
num2words=0.5.12
numba=0.60.0
numpy=1.22.0
nvidia-cublas-cu12=12.1.3.1
nvidia-cuda-cupti-cu12=12.1.105
nvidia-cuda-nvrtc-cu12=12.1.105
nvidia-cuda-runtime-cu12=12.1.105
nvidia-cudnn-cu12=8.9.2.26
nvidia-cufft-cu12=11.0.2.54
nvidia-curand-cu12=10.3.2.106
nvidia-cusolver-cu12=11.4.5.107
nvidia-cusparse-cu12=12.1.0.106
nvidia-nccl-cu12=2.20.5
nvidia-nvjitlink-cu12=12.6.68
nvidia-nvtx-cu12=12.1.105
onnxruntime=1.19.2
openai-whisper=20231117
openssl=3.0.15=h5eee18b_0
orjson=3.10.7
packaging=24.1
pandas=2.0.3
pillow=10.4.0
pip=24.2=py39h06a4308_0
plac=1.4.3
platformdirs=4.3.2
pooch=1.8.2
proces=0.1.7
proto-plus=1.24.0
protobuf=5.28.0
pyasn1=0.6.0
pyasn1-modules=0.4.0
pycparser=2.22
pydantic=2.9.1
pydantic-core=2.23.3
pydub=0.25.1
pygments=2.18.0
pykakasi=2.2.1
pyparsing=3.1.4
pypinyin=0.50.0
python=3.9.19=h955ad1f_1
python-crfsuite=0.9.10
python-dateutil=2.9.0.post0
python-multipart=0.0.9
pytz=2024.1
pyyaml=6.0.2
readline=8.2=h5eee18b_0
referencing=0.35.1
regex=2024.7.24
requests=2.32.3
resampy=0.4.3
rich=13.8.0
rpds-py=0.20.0
rsa=4.9
s3transfer=0.10.2
scikit-learn=1.5.1
scipy=1.11.4
semantic-version=2.10.0
setuptools=72.1.0=py39h06a4308_0
six=1.16.0
sniffio=1.3.1
soundfile=0.12.1
sqlite=3.45.3=h5eee18b_0
starlette=0.38.5
sympy=1.13.2
tensorboard=2.16.2
tensorboard-data-server=0.7.2
threadpoolctl=3.5.0
tiktoken=0.7.0
tk=8.6.14=h39e8969_0
tokenizers=0.13.3
torch=2.3.1
torchaudio=2.3.1
tqdm=4.66.5
transformers=4.27.4
triton=2.3.1
txtsplit=1.0.0
typing-extensions=4.12.2
tzdata=2024.1
tzlocal=5.2
unidecode=1.3.7
unidic=1.1.0
unidic-lite=1.0.8
urllib3=1.26.20
uvicorn=0.30.6
wasabi=0.10.1
wavmark=0.0.3
websockets=11.0.3
werkzeug=3.0.4
wheel=0.43.0
whisper-timestamped=1.14.2
wrapt=1.16.0
xz=5.4.6=h5eee18b_1
zipp=3.20.1
zlib=1.2.13=h5eee18b_1
